refactor: move event handler into component to keep page on server rendering
relates #312
